Move api/testharness to /testharness
I'm not really sure about this, but it seems like a better place than
as a sub package of api.
